Major redesign on API and initial version of BigQuery Sync.
This cset has a major API redesign, and is not backwards
compatible with the previous versions.

This redesign aim to simplify the usage of the aetools package
functions, and to provide a simplified integration with BigQuery tool.

This cset also introduces two new package updates: bigquerysync
and bundle.

On bigquerysync, all basic functionality is present for creating
table schemas and ingesting entities into the datastore.

A new package, bundle, was added to allow a quick usage of the
bigquerysync features, using an alternate app version or module.

--HG--
rename : fixtures.go => funcs.go
rename : fixtures_test.go => funcs_test.go