vmproxy: reuse stopped instances if possible
When attempting to start a new instance, we now detect
if the current instance state is TERMINATED, and if it is,
we then attempt to launch a new one.

Also, we added a timeout for the health check verification,
so if we stop after the backoff factor reaches a minute.

closes #2