aetools: updated to google.golang.org/appengine packages.
The previous version was archived in the v0.1 tag,
if you need it for any historical reasons.
This is a major upgrade, and I'll be releasing
bugfixes as well as security fixes for this newer
version only.

The old `aestubs` package was removed, due to it's
overlap in functionality with https://github.com/luci/gae
It is recomended to migrate to the `gae` package
if you need lightweight, mockable service wrappers.

The `bigquery` package was upgraded to use the
context.Context type, as well as new Google Application
Credentials from golang.org/x/oauth2/google.
As a result, the hacked `serviceaccount` package is
no longer needed and was removed.

The `remoteapi` command, that will be moved away
to a new `cmd` folder in future versions, also uses
the new credential scheme. Support for Client Login
that was used previously was removed. This requires
that the target application use the new configuration
for the remote_api handler, that does not include any
login: admin directive, otherwise an error will happen.